requirements.txt
speechrecognition
pyaudio
pyttsx3
PyPDF2
python-docx
